implemented django backend api call
