+++ frontend/app/build.gradle.kts
@@ -58,12 +58,23 @@ dependencies {
     implementation("com.squareup.retrofit2:retrofit:2.9.0")
     implementation("com.squareup.retrofit2:converter-gson:2.9.0")
     implementation("com.squareup.okhttp3:logging-interceptor:4.9.0")
+    implementation("com.google.code.gson:gson:2.10.1")
     implementation(libs.androidx.core.ktx)
     implementation(libs.androidx.appcompat)
     implementation(libs.material)
     testImplementation(libs.junit)
     androidTestImplementation(libs.androidx.junit)
     androidTestImplementation(libs.androidx.espresso.core)
+}
+
+tasks.register("removePortForwarding") {
+    doLast {
+        exec {
+            executable = "adb"
+            args = listOf("reverse", "--remove", "tcp:8000")
+        }
+        println("Port forwarding removed for port 8000")
+    }
 }
 
 tasks.register("setupPortForwarding") {
@@ -78,6 +89,7 @@ tasks.register("setupPortForwarding") {
 
 tasks.register("installAndRun") {
     dependsOn("installDebug")
+    dependsOn("setupPortForwarding")
     doLast {
         exec {
             executable = "adb"
