+++ frontend/app/src/main/java/com/example/musicroom/MainActivity.kt
@@ -1,5 +1,6 @@
 package com.example.musicroom
 
+import com.example.musicroom.api.RetrofitClient
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
@@ -24,6 +25,7 @@ class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         
+        RetrofitClient.init(this)
         userAuthManager = UserAuthManager(this)
         
         setContent {
