+++ frontend/app/src/main/java/com/example/musicroom/UserAuthManager.kt
@@ -1,35 +1,18 @@
 package com.example.musicroom
 
+import com.example.musicroom.api.*
 import android.content.Context
-import android.content.Intent
-import android.net.Uri
 import android.util.Log
-import androidx.activity.ComponentActivity
-import androidx.activity.result.ActivityResultLauncher
-import androidx.activity.result.contract.ActivityResultContracts
-import androidx.compose.foundation.layout.*
-import androidx.compose.foundation.verticalScroll
-import androidx.compose.foundation.rememberScrollState
-import androidx.compose.foundation.text.KeyboardOptions
-import androidx.compose.material3.*
-import androidx.compose.runtime.*
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.text.input.KeyboardType
-import androidx.compose.ui.text.input.PasswordVisualTransformation
-import androidx.compose.ui.unit.dp
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.withContext
-import kotlinx.coroutines.launch
-import kotlinx.coroutines.MainScope
 import org.json.JSONObject
-import java.net.URL
+import org.json.JSONArray
 
 class UserAuthManager(private val context: Context) {
+    private val TAG = "UserAuthManager"
     private val sharedPreferences = context.getSharedPreferences("user_prefs", Context.MODE_PRIVATE)
     private val _authState = MutableStateFlow<AuthState>(AuthState.NotAuthenticated)
     val authState: StateFlow<AuthState> = _authState.asStateFlow()
@@ -37,9 +20,11 @@ class UserAuthManager(private val context: Context) {
     private val _userProfile = MutableStateFlow<UserProfile?>(null)
     val userProfile: StateFlow<UserProfile?> = _userProfile.asStateFlow()
     
+    private val apiClient = RetrofitClient.api
+    
     init {
-        val isLoggedIn = sharedPreferences.getBoolean("is_authenticated", false)
-        if (isLoggedIn) {
+        val token = sharedPreferences.getString("auth_token", null)
+        if (token != null) {
             loadUserProfile()
             _authState.value = AuthState.Authenticated
         }
@@ -47,8 +32,8 @@ class UserAuthManager(private val context: Context) {
     
     private fun loadUserProfile() {
         val userId = sharedPreferences.getString("user_id", "") ?: ""
+        val username = sharedPreferences.getString("username", "") ?: ""
         val email = sharedPreferences.getString("email", "") ?: ""
-        val displayName = sharedPreferences.getString("display_name", "") ?: ""
         val profilePicUrl = sharedPreferences.getString("profile_pic_url", "") ?: ""
         
         val musicPrefs = sharedPreferences.getString("music_preferences", "[]") ?: "[]"
@@ -56,15 +41,15 @@ class UserAuthManager(private val context: Context) {
         _userProfile.value = UserProfile(
             id = userId,
             email = email,
-            displayName = displayName,
+            displayName = username,
             profilePicUrl = profilePicUrl,
-musicPreferences = parseMusicPrefs(musicPrefs)
+            musicPreferences = parseMusicPrefs(musicPrefs)
         )
     }
     
     private fun parseMusicPrefs(prefsJson: String): List<String> {
         return try {
-            val jsonArray = org.json.JSONArray(prefsJson)
+            val jsonArray = JSONArray(prefsJson)
             List(jsonArray.length()) { i -> jsonArray.getString(i) }
         } catch (e: Exception) {
             emptyList()
@@ -74,7 +59,6 @@ musicPreferences = parseMusicPrefs(musicPrefs)
     suspend fun registerUser(email: String, password: String): Result<Unit> {
         return withContext(Dispatchers.IO) {
             try {
-                
                 if (!android.util.Patterns.EMAIL_ADDRESS.matcher(email).matches()) {
                     return@withContext Result.failure(IllegalArgumentException("Invalid email format"))
                 }
@@ -83,28 +67,46 @@ musicPreferences = parseMusicPrefs(musicPrefs)
                     return@withContext Result.failure(IllegalArgumentException("Password must be at least 8 characters"))
                 }
                 
-                Thread.sleep(1000)
-                
-                val userId = "user_${System.currentTimeMillis()}"
-                sharedPreferences.edit()
-                    .putString("user_id", userId)
-                    .putString("email", email)
-                    .putString("display_name", email.substringBefore("@"))
-                    .putBoolean("is_authenticated", true)
-                    .apply()
-                
-                _authState.value = AuthState.Authenticated
-                _userProfile.value = UserProfile(
-                    id = userId,
+                val username = email.substringBefore("@").replace(".", "_").lowercase()
+                val request = UserRegistrationRequest(
+                    username = username,
                     email = email,
-                    displayName = email.substringBefore("@"),
-                    profilePicUrl = "",
-                    musicPreferences = emptyList()
+                    password = password
                 )
                 
-                Result.success(Unit)
+                val response = apiClient.registerUser(request)
+                
+                if (response.isSuccessful) {
+                    val authResponse = response.body()
+                    if (authResponse != null) {
+                        sharedPreferences.edit()
+                            .putString("auth_token", authResponse.token)
+                            .putString("user_id", authResponse.user.id.toString())
+                            .putString("username", authResponse.user.username)
+                            .putString("email", authResponse.user.email)
+                            .putBoolean("is_authenticated", true)
+                            .apply()
+                        
+                        _authState.value = AuthState.Authenticated
+                        _userProfile.value = UserProfile(
+                            id = authResponse.user.id.toString(),
+                            email = authResponse.user.email,
+                            displayName = authResponse.user.username,
+                            profilePicUrl = "",
+                            musicPreferences = emptyList()
+                        )
+                        
+                        Result.success(Unit)
+                    } else {
+                        Result.failure(Exception("Empty response body"))
+                    }
+                } else {
+                    val errorBody = response.errorBody()?.string() ?: "Unknown error"
+                    Log.e(TAG, "Registration error: $errorBody")
+                    Result.failure(Exception(errorBody))
+                }
             } catch (e: Exception) {
-                Log.e("UserAuthManager", "Error registering user", e)
+                Log.e(TAG, "Error registering user", e)
                 Result.failure(e)
             }
         }
@@ -113,24 +115,81 @@ musicPreferences = parseMusicPrefs(musicPrefs)
     suspend fun loginWithEmail(email: String, password: String): Result<Unit> {
         return withContext(Dispatchers.IO) {
             try {
-                Thread.sleep(1000)
+                val username = email.substringBefore("@")
                 
-                val storedEmail = sharedPreferences.getString("email", null)
-                if (storedEmail != email) {
-                    return@withContext Result.failure(IllegalArgumentException("Email not found"))
+                val request = UserLoginRequest(
+                    username = username,
+                    password = password
+                )
+                
+                val response = apiClient.loginUser(request)
+                
+                if (response.isSuccessful) {
+                    val authResponse = response.body()
+                    if (authResponse != null) {
+                        sharedPreferences.edit()
+                            .putString("auth_token", authResponse.token)
+                            .putString("user_id", authResponse.user.id.toString())
+                            .putString("username", authResponse.user.username)
+                            .putString("email", authResponse.user.email)
+                            .putBoolean("is_authenticated", true)
+                            .apply()
+                        
+                        _authState.value = AuthState.Authenticated
+                        _userProfile.value = UserProfile(
+                            id = authResponse.user.id.toString(),
+                            email = authResponse.user.email,
+                            displayName = authResponse.user.username,
+                            profilePicUrl = "",
+                            musicPreferences = emptyList()
+                        )
+                        
+                        Result.success(Unit)
+                    } else {
+                        Result.failure(Exception("Empty response body"))
+                    }
+                } else {
+                    val errorBody = response.errorBody()?.string() ?: "Unknown error"
+                    Log.e(TAG, "Login error: $errorBody")
+                    Result.failure(Exception(errorBody))
                 }
-                
-                sharedPreferences.edit()
-                    .putBoolean("is_authenticated", true)
-                    .apply()
-                
-                loadUserProfile()
-                
-                _authState.value = AuthState.Authenticated
-                
-                Result.success(Unit)
             } catch (e: Exception) {
-                Log.e("UserAuthManager", "Error logging in", e)
+                Log.e(TAG, "Error logging in", e)
+                Result.failure(e)
+            }
+        }
+    }
+    
+    suspend fun logout(): Result<Unit> {
+        return withContext(Dispatchers.IO) {
+            try {
+                val userProfile = _userProfile.value ?: return@withContext Result.failure(
+                    IllegalStateException("User not logged in")
+                )
+                
+                val request = LogoutRequest(
+                    username = userProfile.displayName
+                )
+                
+                val response = apiClient.logoutUser(request)
+                
+                if (response.isSuccessful) {
+                    sharedPreferences.edit()
+                        .remove("auth_token")
+                        .putBoolean("is_authenticated", false)
+                        .apply()
+                    
+                    _authState.value = AuthState.NotAuthenticated
+                    _userProfile.value = null
+                    
+                    Result.success(Unit)
+                } else {
+                    val errorBody = response.errorBody()?.string() ?: "Unknown error"
+                    Log.e(TAG, "Logout error: $errorBody")
+                    Result.failure(Exception(errorBody))
+                }
+            } catch (e: Exception) {
+                Log.e(TAG, "Error logging out", e)
                 Result.failure(e)
             }
         }
@@ -162,7 +221,6 @@ musicPreferences = parseMusicPrefs(musicPrefs)
     }
     
     fun loginWithFacebook() {
-        
         val userId = "facebook_user_${System.currentTimeMillis()}"
         val email = "facebook_user@example.com"
         val displayName = "Facebook User"
@@ -216,35 +274,24 @@ musicPreferences = parseMusicPrefs(musicPrefs)
                     .putString("public_info", JSONObject(updatedProfile.publicInfo).toString())
                     .putString("friends_only_info", JSONObject(updatedProfile.friendsOnlyInfo).toString())
                     .putString("private_info", JSONObject(updatedProfile.privateInfo).toString())
-                    .putString("music_preferences", JSONObject(updatedProfile.musicPreferences.associateWith { true }).toString())
+                    .putString("music_preferences", JSONArray(updatedProfile.musicPreferences).toString())
                     .apply()
                 
                 _userProfile.value = updatedProfile
                 
                 Result.success(Unit)
             } catch (e: Exception) {
-                Log.e("UserAuthManager", "Error updating profile", e)
+                Log.e(TAG, "Error updating profile", e)
                 Result.failure(e)
             }
         }
     }
     
-    fun logout() {
-        sharedPreferences.edit()
-            .putBoolean("is_authenticated", false)
-            .apply()
-        
-        _authState.value = AuthState.NotAuthenticated
-        _userProfile.value = null
-    }
-    
     fun resetPassword(email: String): Result<Unit> {
-        
         return Result.success(Unit)
     }
     
     fun linkSocialAccount(provider: SocialProvider): Boolean {
-        
         return true
     }
 }
@@ -268,627 +315,3 @@ data class UserProfile(
     val privateInfo: Map<String, String> = emptyMap(),
     val musicPreferences: List<String> = emptyList()
 )
-
-@Composable
-fun LoginScreen(
-    authManager: UserAuthManager,
-    onLoginSuccess: () -> Unit
-) {
-    var email by remember { mutableStateOf("") }
-    var password by remember { mutableStateOf("") }
-    var isLoading by remember { mutableStateOf(false) }
-    var errorMessage by remember { mutableStateOf<String?>(null) }
-    val windowSize = rememberWindowSizeClass()
-    
-    LaunchedEffect(authManager.authState) {
-        authManager.authState.collect { state ->
-            if (state is AuthState.Authenticated) {
-                onLoginSuccess()
-            }
-        }
-    }
-    
-    if (windowSize.isLandscape) {
-        Row(
-            modifier = Modifier
-                .fillMaxSize()
-                .padding(16.dp)
-        ) {
-            Column(
-                modifier = Modifier
-                    .weight(1f)
-                    .fillMaxHeight(),
-                horizontalAlignment = Alignment.CenterHorizontally,
-                verticalArrangement = Arrangement.Center
-            ) {
-                Text(
-                    text = "Music Room",
-                    style = adaptiveTextStyle(windowSize, MaterialTheme.typography.headlineLarge)
-                )
-                
-                Text(
-                    text = "Your collaborative music experience",
-                    style = MaterialTheme.typography.bodyLarge,
-                    modifier = Modifier.padding(top = 8.dp)
-                )
-            }
-            
-            Column(
-                modifier = Modifier
-                    .weight(1f)
-                    .fillMaxHeight()
-                    .verticalScroll(rememberScrollState())
-                    .padding(start = 16.dp),
-                horizontalAlignment = Alignment.CenterHorizontally,
-                verticalArrangement = Arrangement.Center
-            ) {
-                OutlinedTextField(
-                    value = email,
-                    onValueChange = { email = it },
-                    label = { Text("Email") },
-                    modifier = Modifier.fillMaxWidth(),
-                    keyboardOptions = KeyboardOptions(keyboardType = KeyboardType.Email)
-                )
-                
-                Spacer(modifier = Modifier.height(8.dp))
-                
-                OutlinedTextField(
-                    value = password,
-                    onValueChange = { password = it },
-                    label = { Text("Password") },
-                    modifier = Modifier.fillMaxWidth(),
-                    visualTransformation = PasswordVisualTransformation(),
-                    keyboardOptions = KeyboardOptions(keyboardType = KeyboardType.Password)
-                )
-                
-                if (errorMessage != null) {
-                    Text(
-                        text = errorMessage ?: "",
-                        color = MaterialTheme.colorScheme.error,
-                        modifier = Modifier.padding(vertical = 8.dp)
-                    )
-                }
-                
-                Button(
-                    onClick = {
-                        isLoading = true
-                        errorMessage = null
-                        
-                        MainScope().launch {
-                            val result = authManager.loginWithEmail(email, password)
-                            isLoading = false
-                            
-                            result.onFailure { error ->
-                                errorMessage = error.message ?: "Login failed"
-                            }
-                        }
-                    },
-                    modifier = Modifier.fillMaxWidth().padding(vertical = 8.dp),
-                    enabled = !isLoading && email.isNotEmpty() && password.isNotEmpty()
-                ) {
-                    if (isLoading) {
-                        CircularProgressIndicator(
-                            modifier = Modifier.size(24.dp),
-                            color = MaterialTheme.colorScheme.onPrimary
-                        )
-                    } else {
-                        Text("Login")
-                    }
-                }
-                
-                TextButton(onClick = { authManager.resetPassword(email) }) {
-                    Text("Forgot Password?")
-                }
-                
-                Row(
-                    modifier = Modifier.fillMaxWidth().padding(vertical = 8.dp),
-                    horizontalArrangement = Arrangement.SpaceEvenly
-                ) {
-                    Button(
-                        onClick = { authManager.loginWithGoogle() },
-                        colors = ButtonDefaults.buttonColors(
-                            containerColor = MaterialTheme.colorScheme.secondary
-                        )
-                    ) {
-                        Text("Google")
-                    }
-                    
-                    Button(
-                        onClick = { authManager.loginWithFacebook() },
-                        colors = ButtonDefaults.buttonColors(
-                            containerColor = MaterialTheme.colorScheme.tertiary
-                        )
-                    ) {
-                        Text("Facebook")
-                    }
-                }
-                
-                Row(
-                    verticalAlignment = Alignment.CenterVertically,
-                    modifier = Modifier.padding(top = 8.dp)
-                ) {
-                    Text("Don't have an account?")
-                    Spacer(modifier = Modifier.width(4.dp))
-                    TextButton(onClick = {
-                        MainScope().launch {
-                            if (email.isNotEmpty() && password.isNotEmpty()) {
-                                authManager.registerUser(email, password)
-                            } else {
-                                errorMessage = "Please enter email and password to register"
-                            }
-                        }
-                    }) {
-                        Text("Sign Up")
-                    }
-                }
-            }
-        }
-    } else {
-        Column(
-            modifier = Modifier
-                .fillMaxSize()
-                .padding(rememberResponsivePadding(windowSize))
-                .verticalScroll(rememberScrollState()),
-            horizontalAlignment = Alignment.CenterHorizontally,
-            verticalArrangement = Arrangement.Center
-        ) {
-            Text(
-                text = "Music Room",
-                style = adaptiveTextStyle(windowSize, MaterialTheme.typography.headlineMedium)
-            )
-            
-            Spacer(modifier = Modifier.height(24.dp))
-            
-            OutlinedTextField(
-                value = email,
-                onValueChange = { email = it },
-                label = { Text("Email") },
-                modifier = Modifier.fillMaxWidth(),
-                keyboardOptions = KeyboardOptions(keyboardType = KeyboardType.Email)
-            )
-            
-            Spacer(modifier = Modifier.height(8.dp))
-            
-            OutlinedTextField(
-                value = password,
-                onValueChange = { password = it },
-                label = { Text("Password") },
-                modifier = Modifier.fillMaxWidth(),
-                visualTransformation = PasswordVisualTransformation(),
-                keyboardOptions = KeyboardOptions(keyboardType = KeyboardType.Password)
-            )
-            
-            if (errorMessage != null) {
-                Text(
-                    text = errorMessage ?: "",
-                    color = MaterialTheme.colorScheme.error,
-                    modifier = Modifier.padding(vertical = 8.dp)
-                )
-            }
-            
-            Button(
-                onClick = {
-                    isLoading = true
-                    errorMessage = null
-                    
-                    MainScope().launch {
-                        val result = authManager.loginWithEmail(email, password)
-                        isLoading = false
-                        
-                        result.onFailure { error ->
-                            errorMessage = error.message ?: "Login failed"
-                        }
-                    }
-                },
-                modifier = Modifier.fillMaxWidth(),
-                enabled = !isLoading && email.isNotEmpty() && password.isNotEmpty()
-            ) {
-                if (isLoading) {
-                    CircularProgressIndicator(
-                        modifier = Modifier.size(24.dp),
-                        color = MaterialTheme.colorScheme.onPrimary
-                    )
-                } else {
-                    Text("Login")
-                }
-            }
-            
-            TextButton(onClick = { authManager.resetPassword(email) }) {
-                Text("Forgot Password?")
-            }
-            
-            Spacer(modifier = Modifier.height(16.dp))
-            
-            Row(
-                modifier = Modifier.fillMaxWidth(),
-                horizontalArrangement = Arrangement.SpaceEvenly
-            ) {
-                Button(
-                    onClick = { authManager.loginWithGoogle() },
-                    colors = ButtonDefaults.buttonColors(
-                        containerColor = MaterialTheme.colorScheme.secondary
-                    )
-                ) {
-                    Text("Google")
-                }
-                
-                Button(
-                    onClick = { authManager.loginWithFacebook() },
-                    colors = ButtonDefaults.buttonColors(
-                        containerColor = MaterialTheme.colorScheme.tertiary
-                    )
-                ) {
-                    Text("Facebook")
-                }
-            }
-            
-            Spacer(modifier = Modifier.height(16.dp))
-            
-            Row(verticalAlignment = Alignment.CenterVertically) {
-                Text("Don't have an account?")
-                Spacer(modifier = Modifier.width(4.dp))
-                TextButton(onClick = {
-                    MainScope().launch {
-                        if (email.isNotEmpty() && password.isNotEmpty()) {
-                            authManager.registerUser(email, password)
-                        } else {
-                            errorMessage = "Please enter email and password to register"
-                        }
-                    }
-                }) {
-                    Text("Sign Up")
-                }
-            }
-        }
-    }
-}
-
-@Composable
-fun ProfileScreen(
-    authManager: UserAuthManager,
-    onLogout: () -> Unit
-) {
-    val userProfile by authManager.userProfile.collectAsState()
-    val windowSize = rememberWindowSizeClass()
-    
-    var displayName by remember { mutableStateOf(userProfile?.displayName ?: "") }
-    var musicPreference by remember { mutableStateOf("") }
-    var musicPreferences by remember { mutableStateOf(userProfile?.musicPreferences ?: emptyList()) }
-    
-    if (windowSize.isLandscape) {
-        Row(
-            modifier = Modifier
-                .fillMaxSize()
-                .padding(16.dp)
-        ) {
-            Column(
-                modifier = Modifier
-                    .weight(1f)
-                    .fillMaxHeight()
-                    .verticalScroll(rememberScrollState())
-                    .padding(end = 8.dp),
-                horizontalAlignment = Alignment.CenterHorizontally
-            ) {
-                Text(
-                    text = "Your Profile",
-                    style = adaptiveTextStyle(windowSize, MaterialTheme.typography.headlineMedium)
-                )
-                
-                Spacer(modifier = Modifier.height(16.dp))
-                
-                OutlinedTextField(
-                    value = displayName,
-                    onValueChange = { displayName = it },
-                    label = { Text("Display Name") },
-                    modifier = Modifier.fillMaxWidth()
-                )
-                
-                Spacer(modifier = Modifier.height(8.dp))
-                
-                Text(
-                    text = "Email: ${userProfile?.email ?: ""}",
-                    modifier = Modifier.fillMaxWidth()
-                )
-                
-                Spacer(modifier = Modifier.height(16.dp))
-                
-                Text(
-                    text = "Link Accounts",
-                    style = MaterialTheme.typography.titleMedium
-                )
-                
-                Row(
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .padding(vertical = 8.dp),
-                    horizontalArrangement = Arrangement.SpaceEvenly
-                ) {
-                    Button(
-                        onClick = { authManager.linkSocialAccount(SocialProvider.GOOGLE) },
-                        colors = ButtonDefaults.buttonColors(
-                            containerColor = MaterialTheme.colorScheme.secondary
-                        )
-                    ) {
-                        Text("Google")
-                    }
-                    
-                    Button(
-                        onClick = { authManager.linkSocialAccount(SocialProvider.FACEBOOK) },
-                        colors = ButtonDefaults.buttonColors(
-                            containerColor = MaterialTheme.colorScheme.tertiary
-                        )
-                    ) {
-                        Text("Facebook")
-                    }
-                }
-                
-                Spacer(modifier = Modifier.height(16.dp))
-                
-                Button(
-                    onClick = {
-                        MainScope().launch {
-                            authManager.updateProfile(
-                                displayName = displayName
-                            )
-                        }
-                    },
-                    modifier = Modifier.fillMaxWidth()
-                ) {
-                    Text("Save Profile")
-                }
-                
-                Spacer(modifier = Modifier.height(8.dp))
-                
-                Button(
-                    onClick = {
-                        authManager.logout()
-                        onLogout()
-                    },
-                    colors = ButtonDefaults.buttonColors(
-                        containerColor = MaterialTheme.colorScheme.error
-                    ),
-                    modifier = Modifier.fillMaxWidth()
-                ) {
-                    Text("Logout")
-                }
-            }
-            
-            Column(
-                modifier = Modifier
-                    .weight(1f)
-                    .fillMaxHeight()
-                    .verticalScroll(rememberScrollState())
-                    .padding(start = 8.dp),
-                horizontalAlignment = Alignment.CenterHorizontally
-            ) {
-                Text(
-                    text = "Music Preferences",
-                    style = adaptiveTextStyle(windowSize, MaterialTheme.typography.titleMedium),
-                    modifier = Modifier.padding(bottom = 16.dp)
-                )
-                
-                Row(
-                    modifier = Modifier.fillMaxWidth(),
-                    verticalAlignment = Alignment.CenterVertically
-                ) {
-                    OutlinedTextField(
-                        value = musicPreference,
-                        onValueChange = { musicPreference = it },
-                        label = { Text("Add Genre/Artist") },
-                        modifier = Modifier.weight(1f)
-                    )
-                    
-                    Spacer(modifier = Modifier.width(8.dp))
-                    
-                    Button(onClick = {
-                        if (musicPreference.isNotEmpty()) {
-                            musicPreferences = musicPreferences + musicPreference
-                            musicPreference = ""
-                            
-                            MainScope().launch {
-                                authManager.updateProfile(
-                                    musicPreferences = musicPreferences
-                                )
-                            }
-                        }
-                    }) {
-                        Text("Add")
-                    }
-                }
-                
-                Spacer(modifier = Modifier.height(8.dp))
-                
-                musicPreferences.forEach { pref ->
-                    Card(
-                        modifier = Modifier
-                            .fillMaxWidth()
-                            .padding(vertical = 4.dp)
-                    ) {
-                        Row(
-                            modifier = Modifier
-                                .fillMaxWidth()
-                                .padding(8.dp),
-                            horizontalArrangement = Arrangement.SpaceBetween,
-                            verticalAlignment = Alignment.CenterVertically
-                        ) {
-                            Text(pref)
-                            
-                            Button(
-                                onClick = {
-                                    musicPreferences = musicPreferences - pref
-                                    
-                                    MainScope().launch {
-                                        authManager.updateProfile(
-                                            musicPreferences = musicPreferences
-                                        )
-                                    }
-                                },
-                                colors = ButtonDefaults.buttonColors(
-                                    containerColor = MaterialTheme.colorScheme.error
-                                )
-                            ) {
-                                Text("Remove")
-                            }
-                        }
-                    }
-                }
-            }
-        }
-    } else {
-        Column(
-            modifier = Modifier
-                .fillMaxSize()
-                .padding(rememberResponsivePadding(windowSize))
-                .verticalScroll(rememberScrollState()),
-            horizontalAlignment = Alignment.CenterHorizontally
-        ) {
-            Text(
-                text = "Your Profile",
-                style = adaptiveTextStyle(windowSize, MaterialTheme.typography.headlineMedium)
-            )
-            
-            Spacer(modifier = Modifier.height(16.dp))
-            
-            OutlinedTextField(
-                value = displayName,
-                onValueChange = { displayName = it },
-                label = { Text("Display Name") },
-                modifier = Modifier.fillMaxWidth()
-            )
-            
-            Spacer(modifier = Modifier.height(8.dp))
-            
-            Text(
-                text = "Email: ${userProfile?.email ?: ""}",
-                modifier = Modifier.fillMaxWidth()
-            )
-            
-            Spacer(modifier = Modifier.height(16.dp))
-            
-            Text(
-                text = "Music Preferences",
-                style = MaterialTheme.typography.titleMedium
-            )
-            
-            Row(
-                modifier = Modifier.fillMaxWidth(),
-                verticalAlignment = Alignment.CenterVertically
-            ) {
-                OutlinedTextField(
-                    value = musicPreference,
-                    onValueChange = { musicPreference = it },
-                    label = { Text("Add Genre/Artist") },
-                    modifier = Modifier.weight(1f)
-                )
-                
-                Spacer(modifier = Modifier.width(8.dp))
-                
-                Button(onClick = {
-                    if (musicPreference.isNotEmpty()) {
-                        musicPreferences = musicPreferences + musicPreference
-                        musicPreference = ""
-                        
-                        MainScope().launch {
-                            authManager.updateProfile(
-                                musicPreferences = musicPreferences
-                            )
-                        }
-                    }
-                }) {
-                    Text("Add")
-                }
-            }
-            
-            musicPreferences.forEach { pref ->
-                Card(
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .padding(vertical = 4.dp)
-                ) {
-                    Row(
-                        modifier = Modifier
-                            .fillMaxWidth()
-                            .padding(8.dp),
-                        horizontalArrangement = Arrangement.SpaceBetween,
-                        verticalAlignment = Alignment.CenterVertically
-                    ) {
-                        Text(pref)
-                        
-                        Button(
-                            onClick = {
-                                musicPreferences = musicPreferences - pref
-                                
-                                MainScope().launch {
-                                    authManager.updateProfile(
-                                        musicPreferences = musicPreferences
-                                    )
-                                }
-                            },
-                            colors = ButtonDefaults.buttonColors(
-                                containerColor = MaterialTheme.colorScheme.error
-                            )
-                        ) {
-                            Text("Remove")
-                        }
-                    }
-                }
-            }
-            
-            Spacer(modifier = Modifier.height(16.dp))
-            
-            Row(
-                modifier = Modifier.fillMaxWidth(),
-                horizontalArrangement = Arrangement.SpaceEvenly
-            ) {
-                Button(
-                    onClick = { authManager.linkSocialAccount(SocialProvider.GOOGLE) },
-                    colors = ButtonDefaults.buttonColors(
-                        containerColor = MaterialTheme.colorScheme.secondary
-                    )
-                ) {
-                    Text("Link Google")
-                }
-                
-                Button(
-                    onClick = { authManager.linkSocialAccount(SocialProvider.FACEBOOK) },
-                    colors = ButtonDefaults.buttonColors(
-                        containerColor = MaterialTheme.colorScheme.tertiary
-                    )
-                ) {
-                    Text("Link Facebook")
-                }
-            }
-            
-            Spacer(modifier = Modifier.height(16.dp))
-            
-            Button(
-                onClick = {
-                    MainScope().launch {
-                        authManager.updateProfile(
-                            displayName = displayName,
-                            musicPreferences = musicPreferences
-                        )
-                    }
-                },
-                modifier = Modifier.fillMaxWidth()
-            ) {
-                Text("Save Changes")
-            }
-            
-            Spacer(modifier = Modifier.height(8.dp))
-            
-            Button(
-                onClick = {
-                    authManager.logout()
-                    onLogout()
-                },
-                colors = ButtonDefaults.buttonColors(
-                    containerColor = MaterialTheme.colorScheme.error
-                ),
-                modifier = Modifier.fillMaxWidth()
-            ) {
-                Text("Logout")
-            }
-        }
-    }
-}
