+++ frontend/app/src/main/java/com/example/musicroom/api/RetrofitClient.kt
@@ -1,0 +1,42 @@
+package com.example.musicroom.api
+
+import android.content.Context
+import okhttp3.OkHttpClient
+import okhttp3.logging.HttpLoggingInterceptor
+import retrofit2.Retrofit
+import retrofit2.converter.gson.GsonConverterFactory
+import java.util.concurrent.TimeUnit
+
+object RetrofitClient {
+    private const val BASE_URL = "http://localhost:8000/"
+    
+    private lateinit var appContext: Context
+    
+    fun init(context: Context) {
+        appContext = context.applicationContext
+    }
+    
+    private val okHttpClient by lazy {
+        OkHttpClient.Builder()
+            .addInterceptor(HttpLoggingInterceptor().apply {
+                level = HttpLoggingInterceptor.Level.BODY
+            })
+            .addInterceptor(AuthInterceptor(appContext))
+            .connectTimeout(30, TimeUnit.SECONDS)
+            .readTimeout(30, TimeUnit.SECONDS)
+            .writeTimeout(30, TimeUnit.SECONDS)
+            .build()
+    }
+    
+    private val retrofit by lazy {
+        Retrofit.Builder()
+            .baseUrl(BASE_URL)
+            .client(okHttpClient)
+            .addConverterFactory(GsonConverterFactory.create())
+            .build()
+    }
+    
+    val api: MusicRoomApi by lazy {
+        retrofit.create(MusicRoomApi::class.java)
+    }
+}
